requirements.txt
django>=4.2.3,<=4.2.4
djangorestframework
pyyaml
requests
django-cors-headers